Move LSURequester.
